==> src/20_compact_raw.R <==
# This script calculates compactness measures for the maps generated by
# all the algorithm and the existing map.

library(redist)
load("data/data.RData")
load("data/smc.RData")
#load("data/crsg.RData")

# SMC
start.time <- Sys.time()
smc.compact <-
  redist.compactness(
    shp = df,
    district_membership = smc.out$cdvec,
    measure = c("PolsbyPopper", "EdgesRemoved"),
    population = df$pop,
    adjacency = adjlist,
    ncores = 4
  )
end.time <- Sys.time()
print(end.time - start.time)

# # CRSG
# start.time <- Sys.time()
# crsg.compact <-
#   redist.compactness(
#     shp = df,
#     district_membership = crsg.out$partitions,
#     measure = c("PolsbyPopper", "EdgesRemoved"),
#     population = df$pop,
#     adjacency = adjlist,
#     ncores = 4
#   )
# end.time <- Sys.time()
# print(end.time - start.time)
# 
df$SUPDIST <- as.numeric(df$SUPDIST)

# Control
start.time <- Sys.time()
control.compact <-
  redist.compactness(
    shp = df,
    district_membership = df$SUPDIST,
    measure = c("PolsbyPopper", "EdgesRemoved"),
    population = df$pop,
    adjacency = adjlist,
    ncores = 4
  )
end.time <- Sys.time()
print(end.time - start.time)

raw.compact <-
  list(
    smc = smc.compact,
    # crsg = crsg.compact,
    control = control.compact
  )

save(raw.compact, file = "data/compact.raw.RData")
